w3c-reports/12092014/gtk/runned-tests/webmessaging/with-ports/021-expected.txt

Blocked access to external URL http://www1.localhost:8800/webmessaging/without-ports/020-1.html



FAIL cross-origin test assert_equals: expected 2 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/webmessaging/with-ports/021.html:26:52
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
